llm/requirements.txt
langchain
openai
faiss-cpu
tiktoken
python-dotenv
requests
beautifulsoup4
pytube
